Recompiled Samsung Messages with the src code for BlueBubbles
Just so i can run my business more smoothly... I'm always on my mac but despise the iPhone as my daily phone.

With this I can receive and respond messages on my Note 10+ while my iPhone is just a paperweight with a sim card in it -- ONLY because I can answer it on my Mac and iPad as well. I run my own business and got tired of missing calls!